commit changes, removed the dependency on postgres
