+++ src/handlers/collections.rs
@@ -2,27 +2,30 @@ use std::collections::HashMap;
 
 use super::super::facades;
 use axum::extract::{State, Path, Query};
-use axum::response::IntoResponse;
+use axum::{response::IntoResponse, http::{StatusCode, header::{self, HeaderMap, HeaderName}}};
 use deadpool_postgres::{Object, Pool};
 use facades::compression::gzip_compress;
-use facades::postgres_facade::get_offset;
-use facades::efs_facade::{new_read as read_efs, write as write_efs};
+use facades::efs_facade::{get_collection_byte_range as read_efs, append_bytes_collection as write_efs};
 use facades::s3::{read_file as read_s3, init_client as init_s3_client};
 use hyper::body::to_bytes;
-use hyper::{Request, Body, Method, StatusCode};
+use hyper::{Request, Body, Method};
 
 pub async fn collection_handler(
     Path(collection): Path<String>,
-    State(pool_manager): State<Pool>,
     request: Request<Body>
 ) -> impl IntoResponse {
         match *request.method() {
             Method::GET => {
                 let params = extract_query_params(&request.uri().to_string());
-                match (params.get("start").map(|s| s.parse::<i64>()), params.get("end").map(|e| e.parse::<i64>())) {
+                match (params.get("start").map(|s| s.parse::<u64>()), params.get("end").map(|e| e.parse::<u64>())) {
                     (Some(Ok(start)), Some(Ok(end))) => {
                         match get_handler(collection, start, end).await {
-                            Ok(Some(bytes)) => (StatusCode::OK, bytes).into_response(),
+                            Ok(Some(bytes)) => {
+                                let mut headers = HeaderMap::new();
+                                headers.insert(header::CONTENT_TYPE, "application/octet-stream".parse().unwrap());
+                                headers.insert(header::CONTENT_ENCODING, "gzip".parse().unwrap());
+                                (StatusCode::OK, headers, bytes).into_response()
+                            },
                             Ok(None) => (StatusCode::NOT_FOUND, "unable to find supplied byte range".to_string()).into_response(),
                             Err(err) => (StatusCode::INTERNAL_SERVER_ERROR, err).into_response(),
                         }
@@ -31,27 +34,26 @@ pub async fn collection_handler(
                 }
             },
             Method::POST => {
-                let client = pool_manager.get().await.unwrap();
                 let bytes = to_bytes(request.into_body()).await.unwrap().to_vec();
 
-                match post_handler(collection, bytes, client).await{
+                match post_handler(collection, bytes).await{
                     Ok(file_path) => (StatusCode::OK, file_path).into_response(),
                     Err(err) => (StatusCode::INTERNAL_SERVER_ERROR, err).into_response()
                 }
             },
-            Method::PATCH => {
-                let params = extract_query_params(&request.uri().to_string());
-                let bytes = to_bytes(request.into_body()).await.unwrap().to_vec();
-                match (params.get("start").map(|s| s.parse::<i64>()), params.get("end").map(|e| e.parse::<i64>())) {
-                    (Some(Ok(start)), Some(Ok(end))) => {
-                        match patch_handler(collection, bytes, start, end).await {
-                            Ok(file_path) => (StatusCode::OK, file_path).into_response(),
-                            Err(err) => (StatusCode::INTERNAL_SERVER_ERROR, err).into_response()
-                        }
-                    },
-                    _ => (StatusCode::BAD_REQUEST, "unable to parse start & end params".to_string()).into_response()
-                }
-            },
+            // Method::PATCH => {
+            //     let params = extract_query_params(&request.uri().to_string());
+            //     let bytes = to_bytes(request.into_body()).await.unwrap().to_vec();
+            //     match (params.get("start").map(|s| s.parse::<i64>()), params.get("end").map(|e| e.parse::<i64>())) {
+            //         (Some(Ok(start)), Some(Ok(end))) => {
+            //             match patch_handler(collection, bytes, start, end).await {
+            //                 Ok(file_path) => (StatusCode::OK, file_path).into_response(),
+            //                 Err(err) => (StatusCode::INTERNAL_SERVER_ERROR, err).into_response()
+            //             }
+            //         },
+            //         _ => (StatusCode::BAD_REQUEST, "unable to parse start & end params".to_string()).into_response()
+            //     }
+            // },
             _ => StatusCode::NOT_FOUND.into_response()
         }
 }
@@ -62,8 +64,8 @@ pub async fn collection_handler(
 3. Check S3 (return if found)
 4. If nothing found... cry :(
 */
-async fn get_handler(collection: String, start: i64, end: i64) -> Result<Option<Vec<u8>>, String>{
-    let mut res: Option<Vec<u8>> = read_efs(&collection, start, end).await?;
+async fn get_handler(collection: String, start: u64, end: u64) -> Result<Option<Vec<u8>>, String>{
+    let mut res: Option<Vec<u8>> = read_efs(collection, start, end).await?;
 
     if res.is_none() {
         //read S3
@@ -81,33 +83,27 @@ async fn get_handler(collection: String, start: i64, end: i64) -> Result<Option<
 4. Send to EFS
 5. Return file name
 */
-async fn post_handler(collection: String, bytes: Vec<u8>, postgres_client: Object) -> Result<String, String> {
-    //Compress
+async fn post_handler(collection: String, bytes: Vec<u8>) -> Result<String, String> {
     match gzip_compress(bytes) {
         Ok(compressed) => {
-            if let Ok(offsets) = get_offset(postgres_client, collection.clone(), compressed.len()).await {
-                write_efs(compressed, &collection, offsets.0, offsets.1).await
-                    .map(|_|format!("{collection}?start={start}&end={end}",collection = collection, start = offsets.0, end = offsets.1).to_string())
-                    .map_err(|err| err.to_string())
-            } 
-            else {
-                Err("Unable to get offset".to_string())
-            }
+            write_efs(collection, compressed).await
+            .map(|t|format!("{file}?start={start}&end={end}",file = t.0, start = t.1, end = t.2).to_string())
         }
         Err(_) => Err("Unable to compress".to_string()),
     }
 }
-async fn patch_handler(collection: String, bytes: Vec<u8>, start: i64, end: i64) -> Result<String, String> {
-    //Compress
-    match gzip_compress(bytes) {
-        Ok(compressed) => {
-            write_efs(compressed, &collection, start, end).await
-                .map(|_|format!("{collection}?start={start}&end={end}",collection = collection, start = start, end = end).to_string())
-                .map_err(|err| err.to_string())
-        }
-        Err(_) => Err("Unable to compress".to_string()),
-    }
-}
+
+// async fn patch_handler(collection: String, bytes: Vec<u8>, start: i64, end: i64) -> Result<String, String> {
+//     //Compress
+//     match gzip_compress(bytes) {
+//         Ok(compressed) => {
+//             write_efs(compressed, &collection, start, end).await
+//                 .map(|_|format!("{collection}?start={start}&end={end}",collection = collection, start = start, end = end).to_string())
+//                 .map_err(|err| err.to_string())
+//         }
+//         Err(_) => Err("Unable to compress".to_string()),
+//     }
+// }
 
 fn extract_query_params(url: &str) -> HashMap<String, String> {
     let mut params: HashMap<String, String> = HashMap::new();
@@ -130,20 +126,7 @@ mod tests {
     use chrono::{Utc, Datelike};
     use tokio_postgres::Config;
     use super::*;
-    use facades::postgres_facade::{create_config, create_pool};
     use facades::compression::{gzip_decompress as decompress, gzip_compress as compress};
-
-    fn get_test_config() -> Config { create_config("127.0.0.1", "guest", "guest", "test") }
-
-    async fn reset_offset_table(client: Object) -> Result<(), ()> {
-        let query = "DELETE FROM public.\"CacheOffsetTable\" WHERE true;";
-        let statement = client.prepare_cached(&query).await;
-
-        assert!(statement.is_ok());
-
-        client.query(&statement.unwrap(), &[]).await.unwrap();
-        Ok(())
-    }
 
     fn load_test_files() -> Vec<Vec<u8>> {
         let mut files: Vec<Vec<u8>> = Vec::new();
@@ -172,62 +155,23 @@ mod tests {
     }
 
     #[tokio::test]
-    #[ignore = "postgres"]
-    async fn save_test() {
-        let pg_config = get_test_config();
-        let pool = create_pool(pg_config, 1);
-        assert!(pool.is_ok());
-        let client = pool.unwrap().get().await.unwrap();
-
-        let test_base = "UN_FICHIER_TRES_PETIT".to_string();
-        let path = Path::new("test_files/poem.txt");
-        let bytes: Vec<u8> = fs::read(path).unwrap();
-
-        assert!(post_handler(test_base, bytes, client).await.is_ok())
-    }
-
-    #[tokio::test]
-    #[ignore = "postgres"]
-    async fn get_test() {
-        let pg_config = get_test_config();
-        let pool = create_pool(pg_config, 1);
-        assert!(pool.is_ok());
-        let client = pool.unwrap().get().await.unwrap();
-
-        let test_base = "UN_FICHIER_TRES_PETIT".to_string();
-        let path = Path::new("test_files/poem.txt");
-        let bytes: Vec<u8> = fs::read(path).unwrap();
-
-        assert!(post_handler(test_base, bytes, client).await.is_ok())
-    }
-
-    #[tokio::test]
-    #[ignore = "postgres"]
     async fn get_post_integration_test() {
-        let pg_config = get_test_config();
-        let pool = create_pool(pg_config, 1).unwrap();
-        
-        let client = pool.get().await.unwrap();
-        //Make sure we are ready for the test
-        assert!(reset_offset_table(client).await.is_ok());
-
         let test_files = load_test_files();
 
         let test_collection_name = "test_collection".to_string();
 
         for bytes in test_files.iter() {
-            let client = pool.get().await.unwrap();
             let collection_name = test_collection_name.clone();
             //Add the value
-            let post_res = post_handler(collection_name, bytes.to_vec(), client).await;
+            let post_res = post_handler(collection_name, bytes.to_vec()).await;
             assert!(post_res.is_ok());
 
             let file_path = post_res.unwrap();
 
             let params = extract_query_params(&file_path);
 
-            let start = params.get("start").unwrap().parse::<i64>().unwrap();
-            let end = params.get("end").unwrap().parse::<i64>().unwrap();
+            let start = params.get("start").unwrap().parse::<u64>().unwrap();
+            let end = params.get("end").unwrap().parse::<u64>().unwrap();
             let get_res = get_handler(test_collection_name.clone(), start, end).await;
             assert!(get_res.is_ok());
             assert!(get_res.as_ref().unwrap().is_some());
